[作業中] ディレクトリ構成を修正。models モジュールを用意して Reaction クラスはそこに格納。
