+++ src/nasap_net/reaction_pairing/__init__.py
@@ -1,0 +1,2 @@
+from .core import pair_reverse_reactions
+from .models import Reaction
