+++ src/nasap_net/reaction_pairing/models.py
@@ -1,0 +1,15 @@
+from dataclasses import dataclass
+from typing import Generic
+
+from nasap_net.types import A
+
+
+@dataclass(frozen=True)
+class Reaction(Generic[A]):
+    init_assem_id: A
+    entering_assem_id: A | None
+    product_assem_id: A
+    leaving_assem_id: A | None
+    metal_bs: str
+    leaving_bs: str
+    entering_bs: str
