+++ src/nasap_net/types.py
@@ -1,0 +1,5 @@
+from typing import TypeVar
+
+A = TypeVar("A", int, str)  # Assembly ID
+C = TypeVar('C', int, str)  # Component ID
+R = TypeVar('R', int, str)  # Reaction ID
